Change commands to ADD and GET

+++ src/location.h
@@ -2,9 +2,9 @@
 #define REDPILE_COMMON_H
 
 typedef struct {
-    long x;
-    long y;
-    long z;
+    int x;
+    int y;
+    int z;
 } Location;
 
 #define LOCATION_EQUALS(l1,l2) ((l1).x == (l2).x && (l1).y == (l2).y && (l1).z == (l2).z)
